:recycle: Change function location to element
